+++ packages/modal/src/modal.tsx
@@ -24,12 +24,24 @@
  * IN THE SOFTWARE.
  */
 
-import { computed, defineComponent, nextTick, ref, Transition, watch } from 'vue';
+import {
+  computed,
+  defineComponent,
+  type ExtractPropTypes,
+  nextTick,
+  onBeforeUnmount,
+  ref,
+  Transition,
+  watch,
+} from 'vue';
 
 import { usePrefix } from '@bkui-vue/config-provider';
 import { bkZIndexManager, getFullscreenRoot, isElement, isPromise, mask } from '@bkui-vue/shared';
 
+import { useContentResize } from './hooks';
 import { propsMixin } from './props.mixin';
+
+export type ModalProps = Readonly<ExtractPropTypes<typeof propsMixin>>;
 
 export default defineComponent({
   name: 'Modal',
@@ -38,33 +50,31 @@ export default defineComponent({
   },
   emits: ['quick-close', 'quickClose', 'hidden', 'shown', 'close'],
   setup(props, ctx) {
-    const visible = ref(false);
-    const zIndex = ref(props.zIndex);
     const refRoot = ref<HTMLElement>();
     const refMask = ref<HTMLElement>();
+    const teleportTo = ref<string | HTMLElement>('body');
+    const visible = ref(false);
+    const zIndex = ref(props.zIndex);
+    const enableTeleport = ref(!!props.transfer);
     const backgroundColor = ref('rgba(0,0,0,0.6)');
     let closeTimer;
-    const dialogWidth = computed(() => {
-      return /^\d+\.?\d*$/.test(`${props.width}`) ? `${props.width}px` : props.width;
-    });
-    const dialogHeight = computed(() => {
-      return /^\d+\.?\d*$/.test(`${props.height}`) ? `${props.height}px` : props.height;
-    });
 
-    const compStyle = computed(() => {
-      return {
-        width: dialogWidth.value,
-        height: dialogHeight.value,
-        minHeigth: `${200}px`,
+    const { contentStyles, isContentScroll } = useContentResize(refRoot, props);
+
+    const modalWrapperStyles = computed(() => {
+      const baseStyles = {
         display: visible.value ? 'inherit' : 'none',
-        zIndex: zIndex.value || 'inherit',
+      };
+      if (props.fullscreen) {
+        return baseStyles;
+      }
+      return Object.assign(baseStyles, {
+        width: /^\d+\.?\d*$/.test(`${props.width}`) ? `${props.width}px` : props.width,
         left: props.left,
         top: props.top,
         [props.direction]: 0,
-      };
+      });
     });
-    const enableTeleport = ref(!!props.transfer);
-    const teleportTo = ref<string | HTMLElement>('body');
 
     const resolveTransfer = () => {
       if (props.transfer) {
@@ -78,6 +88,7 @@ export default defineComponent({
     };
 
     const { resolveClassName } = usePrefix();
+
     const resolveClosetModal = () => {
       resolveTransfer();
       if (enableTeleport.value) {
@@ -103,12 +114,22 @@ export default defineComponent({
       }
     };
 
-    const fullscreenStyle = computed(() => {
-      return {
-        width: `${100}%`,
-        height: `${100}%`,
-      };
-    });
+    const close = () => {
+      if (visible.value) {
+        visible.value = false;
+        mask.hideMask({
+          el: refRoot.value,
+          mask: refMask.value,
+          showMask: props.showMask,
+          backgroundColor: backgroundColor.value,
+        });
+
+        ctx.emit('hidden');
+        if (enableTeleport.value) {
+          refRoot.value?.remove();
+        }
+      }
+    };
 
     const closeModal = () => {
       mask.hideMask({
@@ -124,7 +145,7 @@ export default defineComponent({
         if (enableTeleport.value) {
           refRoot.value?.remove();
         }
-      }, 250);
+      }, props.hiddenDelay);
     };
 
     watch(
@@ -153,7 +174,9 @@ export default defineComponent({
         visible.value = false;
         closeModal();
       },
-      { immediate: true },
+      {
+        immediate: true,
+      },
     );
 
     const handleBeforeClose = async callbackFn => {
@@ -187,68 +210,76 @@ export default defineComponent({
       }
     };
 
+    onBeforeUnmount(() => {
+      close();
+    });
+
     return {
+      zIndex,
       visible,
-      compStyle,
-      fullscreenStyle,
+      contentStyles,
+      isContentScroll,
+      modalWrapperStyles,
       handleClickOutSide,
       refRoot,
       refMask,
-      showMask: props.showMask,
+      resolveClassName,
+      close,
     };
   },
   render() {
-    const { resolveClassName } = usePrefix();
-    const maxHeight = this.maxHeight ? { maxHeight: this.maxHeight } : {};
-    const bodyClass = `${resolveClassName('modal-body')} ${this.animateType === 'slide' ? this.direction : ''}`;
     return (
       <div
         ref='refRoot'
-        class={[resolveClassName('modal-ctx'), this.visible ? '--show' : '--hide']}
-        style={{ zIndex: this.compStyle.zIndex }}
+        class={{
+          [this.resolveClassName('modal-ctx')]: true,
+          '--show': this.visible,
+          '--hide': !this.visible,
+        }}
+        style={{ zIndex: this.zIndex }}
       >
-        {this.showMask ? (
+        {this.showMask && (
           <div
             ref='refMask'
-            class={[resolveClassName('modal-ctx-mask'), this.visible ? '--show' : '--hide']}
+            class={{
+              [this.resolveClassName('modal-ctx-mask')]: true,
+            }}
             onClick={this.handleClickOutSide}
-            style={{ zIndex: this.compStyle.zIndex }}
-          ></div>
-        ) : (
-          ''
+          />
         )}
-
         <div
-          class={[
-            resolveClassName('modal-wrapper'),
-            this.extCls ?? '',
-            this.bodyClass ?? '',
-            this.size,
-            this.fullscreen ? 'fullscreen' : '',
-          ]}
-          style={[this.compStyle, this.fullscreen ? this.fullscreenStyle : '']}
+          class={{
+            [this.resolveClassName('modal-wrapper')]: true,
+            'scroll-able': this.scrollable,
+            'multi-instance': this.multiInstance,
+          }}
+          style={this.modalWrapperStyles}
         >
           <Transition name={this.animateType}>
-            {this.visible ? (
-              <div class={bodyClass}>
-                <div class={resolveClassName('modal-header')}>{this.$slots.header?.() ?? ''}</div>
-                {this.$slots.default && (
-                  <div
-                    class={resolveClassName('modal-content')}
-                    style={[this.dialogType === 'show' ? 'padding-bottom: 20px' : '', { ...maxHeight }]}
-                  >
-                    {this.$slots.default()}
-                  </div>
-                )}
-                {this.dialogType === 'show' ? (
-                  ''
-                ) : (
-                  <div class={resolveClassName('modal-footer')}>{this.$slots.footer?.() ?? ''}</div>
-                )}
-                {this.closeIcon && <div class={resolveClassName('modal-close')}>{this.$slots.close?.() ?? ''}</div>}
+            {this.visible && (
+              <div
+                class={{
+                  [this.resolveClassName('modal-body')]: true,
+                  [this.direction]: this.animateType === 'slide',
+                }}
+              >
+                <div class={this.resolveClassName('modal-header')}>{this.$slots.header?.()}</div>
+                <div
+                  class={this.resolveClassName('modal-content')}
+                  style={this.contentStyles}
+                >
+                  <div>{this.$slots.default?.()}</div>
+                </div>
+                <div
+                  class={{
+                    [this.resolveClassName('modal-footer')]: true,
+                    'is-fixed': this.isContentScroll,
+                  }}
+                >
+                  {this.$slots.footer?.()}
+                </div>
+                {this.closeIcon && <div class={this.resolveClassName('modal-close')}>{this.$slots.close?.()}</div>}
               </div>
-            ) : (
-              ''
             )}
           </Transition>
         </div>